Remove needless rounding in SDK

+++ src/humanloop/otel/processor.py
@@ -73,10 +73,10 @@ def _process_span_dispatch(span: ReadableSpan, children_spans: list[ReadableSpan
 
     # Processing common to all Humanloop File types
     if span.start_time:
-        span._attributes[f"{HUMANLOOP_LOG_KEY}.start_time"] = int(span.start_time / 1e9)  # type: ignore
+        span._attributes[f"{HUMANLOOP_LOG_KEY}.start_time"] = span.start_time / 1e9  # type: ignore
     if span.end_time:
-        span._attributes[f"{HUMANLOOP_LOG_KEY}.end_time"] = int(span.end_time / 1e9)  # type: ignore
-        span._attributes[f"{HUMANLOOP_LOG_KEY}.created_at"] = int(span.end_time / 1e9)  # type: ignore
+        span._attributes[f"{HUMANLOOP_LOG_KEY}.end_time"] = span.end_time / 1e9  # type: ignore
+        span._attributes[f"{HUMANLOOP_LOG_KEY}.created_at"] = span.end_time / 1e9  # type: ignore
 
     # Processing specific to each Humanloop File type
     if file_type == "prompt":
